+++ CHANGES.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+1.3.0
+-----
+
+* Fix python 3 compatibility
+* Declare support for python 2.7, 3.5, 3.6, 3.7, 3.8
+
 1.2.0
 -----